backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
numpy==1.24.3
tensorflow==2.15.0
onnxruntime==1.16.3
supabase==2.3.0
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.0
